flask_website/requirements.txt
Flask==2.0.1
requests==2.25.1
beautifulsoup4==4.9.3
python-docx==0.8.11